fix: Apply review suggestion.
